Move secretBackendModelConfigWatcher to separate file;
